Move periodic_dynamical_decoupling.py to utils
